backend/requirements.txt
flask>=2.2
flask-cors>=4.0
gunicorn>=21.2
requests>=2.31
beautifulsoup4>=4.12
google-genai>=0.3.0
python-dotenv
gunicorn
firebase-admin>=6.5.0
